dataretriever/requirements.txt
Flask==0.10.1
PyYAML==3.11
Werkzeug==0.10.4
gevent==1.0.1
gunicorn==19.3.0
greenlet==0.4.5
setproctitle==1.1.8
